pyproject.toml
[project]
name = "advanced-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.13",
    "fastapi-mcp>=0.3.4",
    "langchain-openai>=0.3.24",
    "mcp>=1.9.4",
    "python-dotenv>=1.1.0",
    "langchain-community>=0.3.25",
    "langchain-core>=0.3.65",
    "langgraph>=0.4.8",
    "selenium>=4.33.0",
    "undetected-chromedriver>=3.5.5",
    "uvicorn>=0.34.3",
    "langchain-mcp-adapters>=0.1.7",
]
